pyproject.toml
[tool.poetry]
name = "tts-backend"
version = "0.1.0"
description = ""
authors = [
    "Jouko Mälkönen <jouko.malkonen@helsinki.fi>",
    "Soki Tanaka <soki.tanaka@helsinki.fi>",
    "Thomas Grundström <thomas.grundstrom@helsinki.fi>",
    "Toma Lahtinen <toma.lahtinen@helsinki.fi>",
    "Vesa Rautiainen <vesa.rautiainen@helsinki.fi>"
]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "^3.8.1"
django = "^5.1.1"
django-environ = "^0.11.2"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
sqlparse = "^0.5.1"
tzdata = "^2024.2"
pylint-django = "^2.5.5"
invoke = "^2.2.0"

[tool.poetry.group.dev.dependencies]
pylint = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
